refactor(auth): extract shared useAuthFlow hook, kill auth.tsx/web duplication
auth.tsx (native) and auth.web.tsx (web) carried two copies of the same
auth state machine — params, state, the param-sync effect, and the
continue/login/signup/collision/grandfather handlers. Slices 2 and 3 had
to edit both, and one pass missed auth.web.tsx entirely.

Move all of it into components/auth/useAuthFlow.ts. The two screens are
now presentation-only (RN views vs HTML) and consume the same hook, so
behavior can't drift and edits happen once. Heading/subtitle copy is also
centralized; button label stays per-platform ("Log In" vs "Sign In").

Pure refactor — no behavior change. Verified on web: signup + named
applied bar, email-collision flip, full login (email → Welcome back →
Sign In → authenticated), and the "Create one" / "Have an invite?" links.
